Update README.md
HellooooO WORLD, my name is Alejandro, and basically the last time I use Python or even needed to write on english was when i was in the school, so at this point the things that i know about python are equivalent to astrophysics, women or polythics. BUUUT i've always liked to challenge myself, SO what a better way to do it than learning again how to code on python doing it this time fully on english, ANNND obviously in record time. 
Could there be errors? yes 
could there be tears? idk
Will I return to Spanish in the future? probably